feat: major platform enhancements for performance, security, and scalability
Implement several critical improvements to the conversational commerce platform:

- Add optimistic locking with version column to prevent concurrent modification issues
- Implement batch operations for efficient product updates
- Add keyset pagination for better performance with large datasets
- Enhance role-based access control with Clerk token metadata
- Implement audit logging for security-sensitive operations
- Standardize error handling across the API
- Add request timing middleware for performance monitoring
- Refactor Pydantic models to use modern ConfigDict syntax
- Fix 46 deprecation warnings throughout the codebase

These changes significantly improve the platform's readiness for African markets
by enhancing security, performance, and scalability.

+++ backend/alembic/versions/b9e909ee6011_initial_uuid_tables.py
@@ -45,7 +45,7 @@ def upgrade() -> None:
                     sa.Column('id', sa.UUID(), nullable=False),
                     sa.Column('name', sa.String(), nullable=False),
                     sa.Column('description', sa.String(), nullable=False),
-                    sa.Column('price', sa.Float(), nullable=False),
+                    sa.Column('price', sa.Numeric(10, 2), nullable=False),
                     sa.Column('image_url', sa.String(), nullable=True),
                     sa.Column('seller_id', sa.UUID(), nullable=False),
                     sa.Column('created_at', sa.DateTime(), nullable=True),
